PanoDreamer/pyproject.toml
[project]
name = "panodreamer"
version = "0.1.0"
description = "PanoDreamer: Optimization-Based Single Image to 360 3D Scene With Diffusion"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "transformers",
    "diffusers",
    "torch",
    "torchvision",
    "pillow",
    "tqdm",
    "accelerate",
    "kornia",
    "opencv-python",
    "matplotlib",
    "scikit-learn",
    "ropwr",
    "gsplat",
    "gdown",
    "imageio",
    "imageio-ffmpeg",
    "plyfile",
    "einops",
    "huggingface_hub",
    "omegaconf",
    "moviepy<2.0.0",
    "addict",
    "scipy",
    "matplotlib",
    "pycolmap",
    "trimesh",
    "evo",
    "e3nn",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
